sensor/requirements.txt
pyzmq
time
datetime
random
logging